refactor(types): re-export IndirectFireInterfaces from gameplay barrel
Add IndirectFireInterfaces to the src/types/gameplay/index.ts barrel so
existing call sites that import via '@/types/gameplay' are unchanged.

Update 11 direct-import call sites (top-level and inline import()) that
referenced IIndirectFire* / IIndirectFireResolution from CombatInterfaces
to point to the new module.
