+++ src/utils/gameplay/__tests__/declareAttack.indirectFire.test.ts
@@ -17,7 +17,7 @@ import type {
   IIndirectFireResolution,
   IIndirectFireSpotterSelectedPayload,
   IIndirectFireNarcOverridePayload,
-} from '@/types/gameplay/CombatInterfaces';
+} from '@/types/gameplay/IndirectFireInterfaces';
 
 import {
   Facing,
